Address feedback: Move `isRegisteredProperty` and `gentleRegisterProperty` in one file
