+++ src/util.js
@@ -148,25 +148,3 @@ export function splitCommas (value) {
 
 	return ret;
 }
-
-/**
- * Check if a CSS custom property is registered.
- * @param {string} property - The property to check.
- * @param {Window} [window] - The window to check in.
- * @returns {boolean} Whether the property is registered.
- */
-export function isRegisteredProperty (property, window = globalThis) {
-	let document = window.document;
-
-	dummy ??= document.createElement("div");
-	if (!dummy.parentNode) {
-		document.body.append(dummy);
-	}
-
-	let currentValue = getComputedStyle(document.documentElement).getPropertyValue(property);
-	dummy.style.setProperty(property, "foo(bar)"); // a value that is invalid for any registered syntax
-	let newValue = getComputedStyle(dummy).getPropertyValue(property);
-	dummy.remove();
-
-	return currentValue === newValue;
-}
